Allow loading data into memapped numpy arrays
This should address issue #7. Users can specify a directory to be used
to create the memmapped files in when reading large data files. If this
is not specified then data is loaded normally into memory.  Although the
data already exists in the TDMS file, this can't be used by directly by
numpy as the channel data isn't contiguous.