Bugfix of absolute datetime values accuracy.
The previous implementation had a bug that would render inaccuracies
when working with frequencies of which the reciproque had significant
digits < 1e9 seconds, as stated in pullrequest #28.

This could lead to significant errors in the output of larger datasets.

In our example, we had a sampling frequency of 5120 Hertz, which gave an
increment of 0.0001953125 seconds, which rounds to 195312 ns. 3072000
records later, that gives an error of about 1700 us.

I fixed this using the numpy.datetime64, the linspace that was already used
in time_track. I added two parameters to the time_track() function:
  absoluteTime set to True will return numpy.datetime64 objects
  accuracy defines the accuracy of the numpy.datetime64 object.

The accuracy is something that should be handled by the user. Nanosecond
resolution is a good tradeoff between accuracy and possible timespan
(approx 292 years) (limited by the 64 bit datetime datatype of numpy).